More fixes around the madness with linux/ vs net/netinet headers.
Make sure all standard headers were included before including
a linux/ header. In case of netinet/in.h we need to forcefully
include net/if.h becasue both linux/ip6.h and linux/if.h may be
included and the latter blows up if the first is included before
net/if.h.