Make listener_match_peer() work properly in all cases.
The old code only worked if the listener and the peer local-address
where either the same or there was no local-address set.
What did not work was a default listener but a peer with local-address.
Adjust the code and compare using bgpd_addrs to check all the various
options.

This is a fix for one aspect of PR #79